Move advanced orm integration tests into main query file
